agregator/requirements.txt
amqp==2.2.2
asn1crypto==0.23.0
billiard==3.5.0.3
cffi==1.11.0
click==6.7
cryptography==2.0.3
Django==1.11.2
django-bootstrap4==0.0.3
idna==2.6
kombu==4.1.0
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
pycparser==2.18
PyJWT==1.5.3
pytz==2017.2
redis==2.10.6
rq==0.8.2
rq-retry==0.3.0
six==1.11.0
vine==1.1.4
